+++ ext/DiversityPhyloExt.jl
@@ -1,33 +1,22 @@
-using ..Phylo
+module DiversityPhyloExt
+
+isdefined(Base, :get_extension) ? (using Phylo) : (using ..Phylo)
+import Diversity: AbstractPhyloTypes
+import Diversity: PhyloBranches
 using Diversity
 using Diversity.API
 using Statistics
-using AxisArrays
-
-abstract type AbstractPhyloTypes{Tree <: AbstractTree} <:
-    Diversity.API.AbstractTypes
-end
 
 import Diversity.API: _addedoutputcols
 function _addedoutputcols(::AbstractPhyloTypes{TS}) where
-    {LABEL, NL, BL, TS <: TreeSet{LABEL, NL, BL, <: AbstractTree}}
-    Dict{Symbol, Type}(:treename => LABEL)
+    {LABEL, RT, NL, N, B, TS <: TreeSet{LABEL, RT, NL, N, B, <: AbstractTree}}
+    return Dict{Symbol, Type}(:treename => LABEL)
 end
 
 import Diversity.API: _getaddedoutput
 function _getaddedoutput(pt::AbstractPhyloTypes{TS}) where
-    {LABEL, NL, BL, TS <: TreeSet{LABEL, NL, BL, <: AbstractTree}}
-    Dict{Symbol, LABEL}(:treename => first(treenameiter(pt.tree)))
-end
-
-struct PhyloBranches{Tree} <: AbstractPhyloTypes{Tree}
-    tree::Tree
-    nleaf::Int64
-    nancestral::Int64
-    leafnames::Vector{String}
-    ancestralnames::Vector{String}
-    ancestralmatrix::Matrix{Float64}
-    Zmatrix::Matrix{Float64}
+    {LABEL, RT, NL, N, B, TS <: TreeSet{LABEL, RT, NL, N, B, <: AbstractTree}}
+    return Dict{Symbol, LABEL}(:treename => first(gettreenames(pt.tree)))
 end
 
 function PhyloBranches(tree::Tree) where Tree <: AbstractTree
@@ -124,3 +113,5 @@ end
 
 import Diversity.API: _getdiversityname
 _getdiversityname(::PhyloBranches) = "Phylogenetic Branch"
+
+end