+++ test/ext_DiversityPhyloExt.jl
@@ -22,8 +22,18 @@ using Diversity
     @test getabundance(metaphylo, true) ≈ abund
     @test getabundance(metaphylo) ≈ [0.2, 0.2, 0.15, 0.15, 0.3]
     @test getordinariness!(metaphylo) ≈ [0.4, 0.7, 0.3, 0.7, 0.3]
+    @test calcsimilarity(ph, metaphylo.scale) * metaphylo.processedabundances ≈ getordinariness!(metaphylo)
     @test meta_gamma(metaphylo, 0)[!,:diversity] == [2.5]
     @test sub_gamma(metaphylo, 0)[!,:diversity] == [2.5]
+
+    tsph = PhyloBranches(TreeSet(Dict("tree" => nt)))
+    @test species == gettypenames(tsph, true)
+    tsmetaphylo = Metacommunity(abund, tsph)
+    @test species == gettypenames(tsmetaphylo, true)
+    @test meta_gamma(tsmetaphylo, 0).treename == ["tree"]
+    @test subdiv(Gamma(tsmetaphylo), 0).treename == ["tree"]
+    @test metadiv(Gamma(tsmetaphylo), 0).treename == ["tree"]
+    @test all(inddiv(Gamma(tsmetaphylo), 0).treename .== "tree")
 end
 
 end