+++ test/test_Iterators.jl
@@ -1,6 +1,5 @@
 module TestIterators
 using Test
-
 using Phylo
 using Diversity
 
@@ -11,10 +10,28 @@ using Diversity
     abund ./= sum(abund)
     ru = rand(Ultrametric(species))
     m = Metacommunity(abund, PhyloBranches(ru))
-    @test length(TypeIterator(m)) == counttypes(PhyloBranches(ru))
-    @test length(SubcommunityIterator(m)) == sc
-    @test all(reduce(+, SubcommunityIterator(m)) .≈ getmetaabundance(m))
-    @test all(reduce(+, TypeIterator(m)) .≈ getweight(m))
+    ti2 = TypeIterator(m)
+    ti = TypeIterator(getmetaabundance, m)
+    @test length(ti2) == counttypes(PhyloBranches(ru))
+    @test length(ti) == counttypes(PhyloBranches(ru))
+    @test_throws "Can't iterate" TypeIterator(sum ∘ getweight, m)
+
+    si = SubcommunityIterator(m)
+    @test length(si) == sc
+    @test all(reduce(+, si) .≈ getmetaabundance(m))
+    @test all(reduce(+, ti2) .≈ getweight(m))
+    @test all(reduce(+, ti)[1] ≈ 1.0)
+    @test sum(SubcommunityIterator(getweight, m))[1] ≈ 1.0
+    @test_throws "Can't iterate" SubcommunityIterator(sum ∘ getabundance, m)
+
+    @test Base.IteratorSize(typeof(ti)) == Base.HasLength()
+    @test Base.IteratorSize(typeof(ti2)) == Base.HasLength()
+    @test Base.IteratorEltype(typeof(ti)) == Base.HasEltype()
+    @test eltype(ti) ≡ Float64
+
+    @test Base.IteratorSize(typeof(si)) == Base.HasLength()
+    @test Base.IteratorEltype(typeof(si)) == Base.HasEltype()
+    @test eltype(si) ≡ Float64
 end
 
 end