+++ NEWS.md
@@ -1,5 +1,17 @@
 # NEWS
 
+- v0.5.10
+  - Use extensions for Julia 1.9+
+  - Allow AxisArrays to name types
+  - Improve testing
+- v0.5.9
+  - Introduce Gower
+- v0.5.8
+  - Add in Pielou diversity measures
+  - Add docs for PRs
+  - Generalise Jaccard with similarity
+- v0.5.7
+  - Update compat and some doc fixes
 - v0.5.6
   - Fix phylogenetics
 - v0.5.5